Refactor duplicated logger utils/logger.ts
