+++ src/app.ts
@@ -1,11 +1,11 @@
 import express, { Request, Response, NextFunction } from "express";
 import nunjucks from "nunjucks";
 import path from "path";
-import logger from "./lib/Logger";
 import routerDispatch from "./router.dispatch";
 import { servicePathPrefix, ExternalUrls } from "./constants";
 import { sessionMiddleware } from "./middleware/session";
 import cookieParser from "cookie-parser";
+import { logger } from "./lib/Logger";
 
 const app = express();
