+++ src/routers/handlers/confirmCompany/companyNumber.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 import { BaseViewData, GenericHandler, ViewModel } from "./../generic";
 import { ExternalUrls } from "../../../constants";
 import { selectLang } from "../../../utils/localise";
-import logger from "../../../lib/Logger";
+import { logger } from "../../../lib/Logger";
 
 export class CompanyNumberHandler extends GenericHandler<BaseViewData> {
