+++ src/routers/handlers/start/start.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import { BaseViewData, GenericHandler, ViewModel } from "./../generic";
-import logger from "../../../lib/Logger";
+import { logger } from "../../../lib/Logger";
 import { PrefixedUrls } from "../../../constants";
 import { selectLang, getLocalesService, getLocaleInfo } from "../../../utils/localise";
