+++ src/routers/individual.statement.router.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response, Router } from "express";
 import { IndividualStatementHandler } from "./handlers/individual_statement/individual_statement";
 import { handleExceptions } from "../utils/async.handler";
-import logger from "../lib/Logger";
+import { logger } from "../lib/Logger";
 
 const router: Router = Router();
