+++ src/routers/personalCode.router.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response, Router } from "express";
 import { PersonalCodeHandler } from "./handlers/personal_code/personalCode";
 import { handleExceptions } from "../utils/async.handler";
-import logger from "../lib/Logger";
+import { logger } from "../lib/Logger";
 
 const router: Router = Router();
