+++ src/routers/psc.type.router.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response, Router } from "express";
 import { PscTypeHandler } from "./handlers/psc_type/psc_type";
 import { handleExceptions } from "../utils/async.handler";
-import logger from "../lib/Logger";
+import { logger } from "../lib/Logger";
 
 const router: Router = Router();
