+++ src/routers/rlePscListRouter.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response, Router } from "express";
 import { RleListHandler } from "./handlers/rlePscList/rlePscList";
 import { handleExceptions } from "../utils/async.handler";
-import logger from "../lib/Logger";
+import { logger } from "../lib/Logger";
 const router: Router = Router();
 
 router.get("/", handleExceptions(async (req: Request, res: Response) => {
