--- src/utils/logger.ts
@@ -1,10 +1,0 @@
-import { createLogger } from "@companieshouse/structured-logging-node";
-import ApplicationLogger from "@companieshouse/structured-logging-node/lib/ApplicationLogger";
-
-export const logger: ApplicationLogger = createLogger("psc-verification-web");
-
-export const createAndLogError = (description: string): Error => {
-    const error = new Error(description);
-    logger.error(`${error.stack}`);
-    return error;
-};
